Reorganize emoji components
